requirements.txt
telethon>=1.36
python-dotenv>=1.0
MetaTrader5>=5.0.45   # solo Windows
